Move api in accounts to profiles.py for consistency
