+++ backend/app/api/__init__.py
@@ -1,10 +1,9 @@
 """API routers for the Memento application."""
 
-from .accounts import router as accounts_router
 from .consents import router as consents_router
 from .events import router as events_router
 from .memberships import router as memberships_router
-from .profiles import router as profiles_router
+from .profiles import accounts_router, router as profiles_router
 from .recognition import router as recognition_router
 
 __all__ = [
